refactor(annotations): delete the DocumentStore interface and move the wide minter into tests (ADR-035 Unit 8j-1)
The first of three PRs closing Unit 8. This one deletes what has no importer
and relocates the one compatibility export whose confinement was a test
assertion rather than a structural fact. The `ydoc` / `transactMcp` escape
hatches are 8j-2; the doc and ADR corrections are 8j-3.

**The `DocumentStore` interface had zero importers, anywhere.** Not "few" —
zero, in `src/` and `tests/` alike. Every consumer either called
`getDocumentStore`, which returns the concrete `YDocStore`, or constructed the
class directly; the name survived only in docblocks. It was also already dead
as a type: `awareness.ts` reads `store.documentId` three times, and
`documentId` is declared on the class and never on the interface, so returning
`YDocStore` was the only reason that compiled. `EditAnnotationResult` goes the
same way — an alias with no importer, self-referenced three times inside its
own file.

**The wide-typed create wrapper moves to `tests/helpers/annotation-minter.ts`.**
`mcp/annotations.ts::createAnnotation` accepted `note` and `highlight`, which
the seam deliberately refuses, and it had no production caller — only a census
assertion standing between it and acquiring one. It now lives in `tests/`,
where a `src/` file cannot import it. That is the containment the census was
approximating.

**The cost argument for keeping it measured the wrong thing, and the plan said
so before this was written.** The draft cited "174 test call sites across 20
files". 174 is every occurrence of `createAnnotation(` under `tests/` — it
includes the 14 `store.createAnnotation` sites this same PR deletes, and some
`describe` strings. No migration touches a call site: the function is three
lines, an argument reorder and `.id` over `mintAnnotation`, so copying the body
verbatim makes this **16 import-line edits with byte-identical fixtures**. The
parity floor cannot shift, because the body does not change. `tests/helpers/
positions.ts` already documents this exact move for a ~294-site problem: "one
import per file rather than a cast per call."

`mintAnnotation` stays in `lifecycle.ts` — it performs the real origin-tagged
Y.Map write and fires the notification, so it is production code with a
test-only caller. Its census assertion flips from "exactly one importer" to
"none", and carries a positive anchor so the empty set is not a zero check.

**`acceptPending` / `dismissPending` were considered for the same move and
declined, with a reason.** They delegate to `transitionPending`, which is
module-private: relocating them means exporting the guarded transition
mechanism itself, trading two narrow test-only exports for one wide one. The
census pinning their `src/` importer set to `[]` is the cheaper containment.
Recorded in the docblock so the next reader does not re-derive it.

**Two specs are deleted rather than ported.** The `YDocStore.createAnnotation`
parity pair pinned a method with no production callers, and its origin spec is
strictly weaker than what already exists: it captured the last `afterTransaction`
origin, where `annotation-create-lifecycle.test.ts:151` asserts
`toStrictEqual([MCP_ORIGIN])` — exactly one transaction. The `getAnnotation`
spec was named "returns the sanitized record" and asserted only `.content` on a
record it had just minted through the canonical path; it never built a legacy
shape and never touched the relay, so it pinned presence and absence, not
sanitization. Everything else that called those two methods was setup and moves
to a local `mint` shim writing the same record through the same function into
the same map.

**Two census expectations narrow, and the narrowing is the point.** The
`createAnnotation(` sweep drops from three files to one — and the survivor,
`Toolbar.svelte`, is a *different* function: the client's own local declaration,
writing under `withBrowser`. It stays unfiltered, both because the sweep is
deliberately wider than the thing it guards and because it is what keeps the
assertion from being a zero check.

**One deletion in this PR was wrong and the compiler caught it.** `filePath`
was reported by two independent census passes as having zero readers anywhere.
It has one: `mcp/annotations.ts:632` binds it in a DESTRUCTURE —
`const { ydoc, filePath } = store` — which no member-name grep sees, and which
feeds four uses in `tandem_exportAnnotations`. The field is restored with the
reason recorded on it. That line is also a fifth production use of the `ydoc`
hatch that every census of it has missed, including one that explicitly checked
for destructures and searched for the single-name form. 8j-2 inherits it.

Five test files used a multi-line import block the first migration pass could
not see for the same reason — a line-shaped sweep against a symbol reached
through another shape. `typecheck:tests` (#1616) is what caught those.

Verified: `npm run typecheck`, `npm run typecheck:tests`, biome and
`audit:origins` clean; `tests/server/` green at 4221 passed / 27 skipped.

Refs #315
