+++ src/server/mcp/document-store.ts
@@ -1,20 +1,19 @@
 /**
- * DocumentStore — a testable seam over the in-memory Y.Doc / Y.Map layer used
- * by the MCP tool handlers (issue #315).
+ * `YDocStore` — a testable seam over the in-memory Y.Doc / Y.Map layer used by
+ * the MCP tool handlers (issue #315).
  *
  * Before this seam, every annotation/text tool reached straight into a raw
- * `Y.Doc` + `Y.Map` (via `getDocAndAnnotations`). That coupled the tool logic
- * to Yjs internals and made the handlers awkward to test without standing up a
- * full document service. `DocumentStore` names the operations the handlers
- * actually perform — read text, create/list/edit/resolve/remove annotations,
- * add/list replies, refresh CRDT ranges — so handlers depend on an interface
- * instead of `Y.Map.get`/`set`.
+ * `Y.Doc` + `Y.Map` (via `getDocAndAnnotations`, which handed back the Y.Map
+ * itself). That coupled the tool logic to Yjs internals and made the handlers
+ * awkward to test without standing up a full document service. This class names
+ * the operations the handlers actually perform — read text, list/edit/resolve/
+ * remove annotations, add/list replies, refresh CRDT ranges — so handlers never
+ * write `Y.Map.get`/`set`.
  *
- * `YDocStore` is the one implementation. It is intentionally a thin wrapper:
- * it delegates to the same standalone helpers the handlers used before
- * (`createAnnotation`, `collectAnnotations`, `refreshAllRanges`)
- * and, for the families that have
- * migrated, to {@link DocumentStore.lifecycle}. That delegation is the parity
+ * `YDocStore` is the one implementation, and as of Unit 8j the only type: it
+ * delegates every mutation to {@link YDocStore.lifecycle} and its remaining
+ * reads to the same standalone helpers the handlers used before
+ * (`collectAnnotations`, `refreshAllRanges`). That delegation is the parity
  * contract: the underlying Y.Map structures and origin tagging (`withMcp`,
  * ADR-031) are byte-identical to the pre-refactor behavior. The helpers stay
  * exported because the HTTP routes and the existing test suite (the parity
@@ -22,27 +21,24 @@
  *
  * **Direction of travel (ADR-035).** This store is a compatibility shell, not
  * the destination. Unit 8b settled that `AnnotationLifecycle` is the seam
- * callers program against and that Unit 8j collapses or deletes this file. Two
- * consequences for anyone editing here: new mutation work goes on the
- * lifecycle, not on a new method here; and `ydoc` / `transactMcp` are the
- * escape hatches the Unit 8 epic in
+ * callers program against. Two consequences for anyone editing here: new
+ * mutation work goes on the lifecycle, not on a new method here; and `ydoc` /
+ * `transactMcp` are the escape hatches the Unit 8 epic in
  * `docs/plans/2026-08-24-ai-assisted-maintainability-remediation.md` exists to
  * remove (ADR-035 predates this store and never names it), so nothing new
- * should reach for them.
+ * should reach for them. **Unit 8j-2 closes them; until it lands they are
+ * still here and still reachable.**
  *
  * Scope note: this wraps the *in-memory* Y.Doc/Y.Map layer the MCP handlers
  * touch — NOT the durable annotation file-store (`src/server/annotations/`).
- * `FileOnlyStore` is intentionally out of scope; only the interface +
- * `YDocStore` ship here.
+ * `FileOnlyStore` is intentionally out of scope.
  */
 
 import type * as Y from "yjs";
 import { Y_MAP_ANNOTATION_REPLIES, Y_MAP_ANNOTATIONS } from "../../shared/constants.js";
 import { withMcp } from "../../shared/origins.js";
-import type { AnchoredRangeResult } from "../../shared/positions/index.js";
 import type { SanitizationEvent } from "../../shared/sanitize.js";
-import { sanitizeAnnotation } from "../../shared/sanitize.js";
-import type { Annotation, AnnotationReply, AnnotationType } from "../../shared/types.js";
+import type { Annotation, AnnotationReply } from "../../shared/types.js";
 import { docHash } from "../annotations/doc-hash.js";
 import {
   type AnnotationLifecycle,
@@ -51,171 +47,38 @@ import {
   type EditPatch,
   type EditResult,
   type LifecycleResult,
-  type MintExtras,
   type RemoveResult,
 } from "../annotations/lifecycle.js";
 import { relaySanitizationEvent } from "../annotations/migration-log.js";
 import { refreshAllRanges, refreshRange } from "../positions.js";
 import { getOrCreateDocument } from "../yjs/provider.js";
-import {
-  collectAnnotations,
-  collectRepliesForAnnotation,
-  createAnnotation,
-} from "./annotations.js";
+import { collectAnnotations, collectRepliesForAnnotation } from "./annotations.js";
 import { extractText } from "./document-model.js";
 import { getCurrentDoc } from "./document-service.js";
 
 /**
- * Operations the MCP tool handlers perform against a single document's
- * in-memory Y.Doc / Y.Map state. Methods mirror the handler call sites:
- * the names and argument shapes are derived from what the handlers do, not
- * from the underlying Y.Map API.
+ * The operations MCP tool handlers perform against a single document's
+ * in-memory Y.Doc / Y.Map state. Method names and argument shapes are derived
+ * from what the handlers do, not from the underlying Y.Map API.
+ *
+ * **There is no `DocumentStore` interface any more (Unit 8j).** It had zero
+ * importers anywhere in `src/` or `tests/` — every consumer either called
+ * {@link getDocumentStore}, which returns this class, or constructed it
+ * directly. It was also already dead as a type: `awareness.ts` reads
+ * `store.documentId`, declared here and never on the interface, so returning
+ * the concrete class was the only reason that compiled. A one-implementation
+ * interface nothing imports is documentation wearing a type's clothes.
  */
-export interface DocumentStore {
-  /** Underlying Y.Doc — escape hatch for range anchoring and text extraction. */
+export class YDocStore {
   readonly ydoc: Y.Doc;
   /**
-   * The ADR-035 mutation seam for this document (Unit 8b).
+   * Absolute (or `upload://`) path of the backing document.
    *
-   * **This is the interface callers program against from here on; the store is
-   * a compatibility shell Unit 8j deletes.** It is reached through the store
-   * only because document *resolution* still lives here — moving
-   * `getCurrentDoc` / `getOrCreateDocument` into `annotations/lifecycle.ts`
-   * would make `annotations/ → mcp/` an import cycle. 8j moves the lookup.
-   *
-   * Per the lifecycle's lifetime rule this is safe only because
-   * {@link getDocumentStore} constructs a fresh store per handler call. Do not
-   * cache a store — or this lifecycle — across an `await` that could span a
-   * Hocuspocus `onLoadDocument` doc swap.
+   * Read by `tandem_exportAnnotations` (`mcp/annotations.ts:632`) through a
+   * DESTRUCTURE — `const { ydoc, filePath } = store` — which is why two
+   * independent census passes over this field reported zero readers and Unit 8j
+   * briefly deleted it. A member-name grep does not see a destructured bind.
    */
-  readonly lifecycle: AnnotationLifecycle;
-  /** Absolute (or `upload://`) path of the backing document. */
-  readonly filePath: string;
-  /** Stable hash of `filePath`, used to key migration-log relays. */
-  readonly docHash: string;
-
-  // --- Text ---
-
-  /**
-   * Full document text in the annotation coordinate system (flat offsets,
-   * heading prefixes included). Always `extractText`, never `extractMarkdown`
-   * (Critical Rule #5).
-   */
-  getText(): string;
-
-  // --- Annotations: write ---
-
-  /**
-   * Create a Claude-authored annotation from an already-anchored range and
-   * store it in the annotations Y.Map. Returns the new annotation ID.
-   *
-   * @deprecated ADR-035 Unit 8b — use {@link DocumentStore.lifecycle}`.create`.
-   * It has **no production callers left**: what survives is the test floor,
-   * which builds `note` and `highlight` fixtures through it and which the seam
-   * deliberately cannot express. (The not-yet-migrated families, Units 8c–8h,
-   * do not reach for this method either — `.docx` import has its own
-   * `withInternal` path.) Unit 8j removes it.
-   */
-  createAnnotation(
-    type: AnnotationType,
-    anchored: AnchoredRangeResult,
-    content: string,
-    extras?: MintExtras,
-  ): string;
-
-  /**
-   * Edit the mutable fields of a pending annotation. Returns the updated
-   * record, or a tagged failure (`not-found` / `invalid-note` /
-   * `not-pending` / `empty-patch` / `invalid-suggestion-target`) so the
-   * handler can map it to the right MCP error envelope. The failure-arm
-   * order mirrors the pre-seam handler's sequential guards exactly.
-   */
-  editAnnotation(id: string, patch: EditPatch): EditAnnotationResult;
-
-  /** Accept a pending annotation (pending → accepted). */
-  acceptAnnotation(id: string): LifecycleResult<Annotation>;
-  /** Dismiss a pending annotation (pending → dismissed). */
-  dismissAnnotation(id: string): LifecycleResult<Annotation>;
-
-  /** Remove an annotation and its orphaned replies, on CLAUDE's behalf. */
-  removeAnnotation(id: string): RemoveResult;
-
-  // --- Annotations: read ---
-
-  /** Get a single annotation by ID (sanitized), or undefined if absent. */
-  getAnnotation(id: string): Annotation | undefined;
-
-  /** Collect all annotations as a sanitized array (skips malformed rows). */
-  listAnnotations(): Annotation[];
-
-  /**
-   * Collect annotations and refresh their CRDT-anchored ranges in one pass,
-   * persisting any range updates back to the Y.Map. Returns the refreshed
-   * annotations.
-   */
-  listAnnotationsRefreshed(): Annotation[];
-
-  /**
-   * Refresh a single annotation's CRDT-anchored range, persisting any update
-   * back to the Y.Map, and return the refreshed annotation. Used by the inbox
-   * surfacer, which refreshes a surfaced-gated subset rather than the whole
-   * collection. The caller is responsible for the enclosing origin-tagged
-   * transaction (see {@link DocumentStore.transactMcp}).
-   */
-  refreshAnnotation(ann: Annotation): Annotation;
-
-  /** Run `fn` inside an MCP-origin Y.Doc transaction (ADR-031 `withMcp`). */
-  transactMcp(fn: () => void): void;
-
-  // --- Replies ---
-
-  /**
-   * Add CLAUDE's reply to a comment thread, carrying the ADR-027 guard.
-   *
-   * **No `author` parameter, and its absence is the point.** It took
-   * `ReplyAuthor` — three members — so `store.addReply(id, text, "import")` was
-   * type-legal. Master wrote that value straight through as `author: "import"`;
-   * what it bought was skipping the `author === "claude"` guard entirely, so
-   * Claude could reply into a note thread by picking a third byline the client
-   * renders as an import. (An earlier draft of this sentence said it "would have
-   * been written as a USER reply" — a more specific claim, and the wrong one:
-   * `author` was never remapped, and `heldInSolo` is gated on `"user"`, so an
-   * `"import"` reply would not have been stamped either.) Keeping the parameter
-   * would also have re-keyed the privacy guard on a
-   * caller-supplied value one level above the seam, which is the defect Unit 8f
-   * exists to remove. This store is the MCP surface; the browser's entry is
-   * `addUserReply`.
-   */
-  addReply(annotationId: string, text: string): ClaudeReplyResult;
-
-  /** Collect all replies for an annotation, sorted chronologically. */
-  listReplies(annotationId: string): AnnotationReply[];
-}
-
-/**
- * Tagged outcome of {@link DocumentStore.editAnnotation}.
- *
- * @deprecated Use `EditResult` from `annotations/lifecycle.ts`.
- *
- * An ALIAS, not a copy. Restating the union here would leave two structurally
- * identical types free to drift, which is the failure the lifecycle module
- * exists to prevent.
- *
- * It survives only as a name. Review measured that the MCP adapter never
- * mentions it — `mcp/annotations.ts` switches on `result.kind` off an
- * unannotated `store.editAnnotation(...)` — so this alias is not load-bearing
- * for any consumer, and Unit 8j deletes it.
- */
-export type EditAnnotationResult = EditResult;
-
-/**
- * The lone {@link DocumentStore} implementation. Wraps a document's Y.Doc and
- * its annotations Y.Map, delegating every mutation to the existing helpers so
- * the Y.Map structures and origin tagging are unchanged from the pre-seam
- * handlers.
- */
-export class YDocStore implements DocumentStore {
-  readonly ydoc: Y.Doc;
   readonly filePath: string;
   readonly docHash: string;
   /**
@@ -249,16 +112,6 @@ export class YDocStore implements DocumentStore {
     return extractText(this.ydoc);
   }
 
-  /** @deprecated see {@link DocumentStore.createAnnotation}. */
-  createAnnotation(
-    type: AnnotationType,
-    anchored: AnchoredRangeResult,
-    content: string,
-    extras?: MintExtras,
-  ): string {
-    return createAnnotation(this.map, this.ydoc, type, anchored, content, extras);
-  }
-
   /**
    * Delegates to the lifecycle (ADR-034/035 Unit 8c). The guards, the `rev`
    * bump and the `withMcp` tag all live there now.
@@ -266,7 +119,7 @@ export class YDocStore implements DocumentStore {
    * **The sink is passed, not defaulted**, and as of Unit 8d every mutation
    * method on this store does the same.
    */
-  editAnnotation(id: string, patch: EditPatch): EditAnnotationResult {
+  editAnnotation(id: string, patch: EditPatch): EditResult {
     return this.lifecycle.editPending(id, patch, (e) => this.onLossy(e));
   }
 
@@ -307,12 +160,6 @@ export class YDocStore implements DocumentStore {
    */
   removeAnnotation(id: string): RemoveResult {
     return this.lifecycle.remove(id, (e) => this.onLossy(e));
-  }
-
-  getAnnotation(id: string): Annotation | undefined {
-    const raw = this.map.get(id) as Annotation | undefined;
-    if (!raw) return undefined;
-    return sanitizeAnnotation(raw, (e) => this.onLossy(e));
   }
 
   listAnnotations(): Annotation[] {
