+++ tests/server/annotation-remove-seam.test.ts
@@ -26,13 +26,13 @@ import type { Request, Response } from "express";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import { addUserReply, removeAnnotationRecord } from "../../src/server/annotations/lifecycle.js";
 import { makeAnnotationsObserver } from "../../src/server/events/observers/annotations.js";
-import { createAnnotation } from "../../src/server/mcp/annotations.js";
 import { YDocStore } from "../../src/server/mcp/document-store.js";
 import { handleRemoveAnnotation } from "../../src/server/mcp/routes/remove-annotation.js";
 import { getBuffer, resetForTesting } from "../../src/server/notifications.js";
 import { Y_MAP_ANNOTATION_REPLIES, Y_MAP_ANNOTATIONS } from "../../src/shared/constants.js";
 import type { TandemEvent } from "../../src/shared/events/types.js";
 import { BROWSER_ORIGIN, MCP_ORIGIN, withBrowser } from "../../src/shared/origins.js";
+import { createAnnotation } from "../helpers/annotation-minter.js";
 import { clearOpenDocs, setupDoc } from "../helpers/doc-service.js";
 import { unanchored } from "../helpers/positions.js";
 import { filesMentioning, SRC_FILES, stripComments } from "../helpers/src-tree.js";
