+++ tests/server/mcp-output-schemas.test.ts
@@ -35,7 +35,7 @@ import { beforeEach, describe, expect, it } from "vitest";
 import { type ZodRawShape, z } from "zod";
 import type { DoctorReport } from "../../src/cli/doctor.js";
 import { addDoc, removeDoc, setActiveDocId } from "../../src/server/documents/registry-testing.js";
-import { createAnnotation, registerAnnotationTools } from "../../src/server/mcp/annotations.js";
+import { registerAnnotationTools } from "../../src/server/mcp/annotations.js";
 import { registerAwarenessTools, resetInbox } from "../../src/server/mcp/awareness.js";
 import { registerDiagnosticsTools } from "../../src/server/mcp/diagnostics.js";
 import { populateYDoc, registerDocumentTools } from "../../src/server/mcp/document.js";
@@ -56,6 +56,7 @@ import { CTRL_ROOM, Y_MAP_ANNOTATIONS, Y_MAP_CHAT } from "../../src/shared/const
 import { withInternal } from "../../src/shared/origins.js";
 import type { Annotation, ChatMessage } from "../../src/shared/types.js";
 import { toFlatOffset } from "../../src/shared/types.js";
+import { createAnnotation } from "../helpers/annotation-minter.js";
 import { rangeOf } from "../helpers/ydoc-factory.js";
 
 let client: Client;
