+++ tests/server/remove-annotation.test.ts
@@ -14,8 +14,8 @@ import {
   registerAnnotationObserver,
   resetForTesting,
 } from "../../src/server/annotations/sync.js";
-import { createAnnotation } from "../../src/server/mcp/annotations.js";
 import { Y_MAP_ANNOTATION_REPLIES, Y_MAP_ANNOTATIONS } from "../../src/shared/constants.js";
+import { createAnnotation } from "../helpers/annotation-minter.js";
 import { useTmpAnnotationsEnvWithFlag } from "../helpers/annotation-store-env.js";
 import { clearOpenDocs, setupDoc } from "../helpers/doc-service.js";
 import { unanchored } from "../helpers/positions.js";
