+++ tests/server/resolve-annotation.test.ts
@@ -12,13 +12,14 @@ import { InMemoryTransport } from "@modelcontextprotocol/sdk/inMemory.js";
 import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { beforeEach, describe, expect, it } from "vitest";
 import { addDoc, removeDoc, setActiveDocId } from "../../src/server/documents/registry-testing.js";
-import { createAnnotation, registerAnnotationTools } from "../../src/server/mcp/annotations.js";
+import { registerAnnotationTools } from "../../src/server/mcp/annotations.js";
 import { populateYDoc } from "../../src/server/mcp/document.js";
 import { getOpenDocs } from "../../src/server/mcp/document-service.js";
 import { getOrCreateDocument } from "../../src/server/yjs/provider.js";
 import { Y_MAP_ANNOTATIONS } from "../../src/shared/constants.js";
 import { MCP_ORIGIN } from "../../src/shared/origins.js";
 import type { Annotation } from "../../src/shared/types.js";
+import { createAnnotation } from "../helpers/annotation-minter.js";
 import { rangeOf } from "../helpers/ydoc-factory.js";
 
 let client: Client;
